==> 2.a_Census_LFS_mapping.Rmd <==
---
title: "Mapping LFS to Census Indutries"
author: "Sandrine Njoo"
date: "25/10/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Set up the analytic environment
Load libraries.
```{r library, warning = FALSE, message = FALSE}
#using R:

library(tidyverse)  # data manipulation and piping
library(data.table) # renaming columns in bulk using 'setnames'
library(plyr)       # recoding factor levels using 'mapvalues'  

# set up data location:
location = "C:/Users/wher/OneDrive - Bank of Canada/MyDocs/git"

```
R version used for this analysis

```{r}
version
```



## Reading the LFS and the Census data 
Extracting the the industry names for both datasets 

```{r lfsload, warning = FALSE, message = FALSE}
# Use of just one month / one province is just for speed, rather than loading the full country, for example: '1.b-lfs_data_recoded_vars-from-2021-09.csv'
lfs <- read_csv(paste0(location,"1.b-lfs_recoded_vars-Prince_Edward_Island-2021-09.csv"))
lfs_values <- unique(lfs$NAICS_21.INDUSTRY_OF_MAIN_JOB)
lfs_values = lfs_values[!is.na(lfs_values)]
lfs_values
```

```{r, warning = FALSE, message = FALSE}

# province choice should be arbitrary, but we need to load a province
Province = "PE" # two digit province name code.

census_data        <- read_csv(paste0(location,"1.a_census_data_",Province,".csv"))
census_data_values <- grep(colnames(census_data), pattern = "^X", value = TRUE)

# remove variables that were previously merged into new broader categories: 
census_data_values <- setdiff(census_data_values, c("X11.Agriculture..forestry..fishing.and.hunting",
                                                    "X21.Mining..quarrying..and.oil.and.gas.extraction",
                                                    "X41.Wholesale.trade","X44.45.Retail.trade",
                                                    "X52.Finance.and.insurance","X53.Real.estate.and.rental.and.leasing",
                                                    "X55.Management.of.companies.and.enterprises","X56.Administrative.and.support..waste.management.and.remediation.services",
                                                    "X51.Information.and.cultural.industries",
                                                    "X71.Arts..entertainment.and.recreation"))


census_data_values
```

## Creating a dictionary

Maps the industry from the LFS to the Census data using approximate string matching.  Some strings will match easily.  Others will match more readily if the punctuation is altered first by replacing "_" with ".".
```{r}
dict_new <- data.frame(lfs_data=NA, census=NA)[numeric(0), ]
not_matched <- data.frame(not_found=NA)

for (val in 1:length(lfs_values)){
  if(!identical(agrep(lfs_values[val], census_data_values), integer(0)) | # try as is
     !identical(agrep(gsub(lfs_values[val],pattern="_",replacement="."), census_data_values), integer(0))){ # look for approximate match with a few punctuation changes
    out_1 = lfs_values[val]
    
    # catch if we need to replace punctuation
    if(!identical(agrep(lfs_values[val], census_data_values), integer(0))){
      out_2 = census_data_values[agrep(lfs_values[val], census_data_values)]
    }else{# look for approximate match with a few punctuation changes
      out_2 = census_data_values[agrep(gsub(lfs_values[val],pattern="_",replacement="."), census_data_values)]
    }
    if(length(out_2)<2){
      dict_new[val,1] <- out_1
      dict_new[val,2] <- out_2
    }
    else{
      if(adist(lfs_values[val],out_2[1])>adist(lfs_values[val],out_2[2])){
        dict_new[val,1] <- out_1
        dict_new[val,2]<-out_2[2]
      }
      else{
        dict_new[val,2]<-out_2[1]
      }
    }
  }
  else if(identical(agrep(lfs_values[val], census_data_values), integer(0))){
    out = lfs_values[val]
    not_matched = rbind(not_matched, out)
  }
  
}
dict_new <- dict_new %>% drop_na()


  
```

Manually adding the values that were not matched.  These are stored in the **not_matched** variable.
```{r}
# handle these cases manually:
not_matched

dict_new[nrow(dict_new) + 1,] = c("Fishing__hunting_and_trapping","X11.Agriculture.forestry.fishing.hunting.21.Mining.quarrying.and.oil.and.gas.extraction")
dict_new[nrow(dict_new) + 1,] = c("Manufacturing___non_durable_goods","X31.33.Manufacturing")
dict_new[nrow(dict_new) + 1,] = c("Manufacturing___durable_goods","X31.33.Manufacturing")
dict_new[nrow(dict_new) + 1,] = c("Forestry_and_logging_and_support_activities_for_forestry","X11.Agriculture.forestry.fishing.hunting.21.Mining.quarrying.and.oil.and.gas.extraction")
```


Display and save the dictionary
```{r}
dict_new

dict_new %>% write_csv(file = paste0(location,"2.a-LFT_to_census_dictionary.csv"))
```
